requirements.txt
pytest
numpy
seaborn
matplotlib
fire
pandas
Pillow
imageio
pygifsicle
